# api/openapi-spec/openapi.yaml
openapi: 3.0.3
info:
  title: Attendee Service
  description: |-
    This backend microservice manages attendee registrations. This includes    
    - creating and editing registrations
    - adding and removing packages incl. calculating total due amount (but not current dues)
    - searching and listing registrations (admin only)
    - registration status progression (new - approved - partially paid - paid - checked in / waiting / cancelled / deleted)
    - ban detection and management
    - transferring registrations when re-sold
  license:
    name: MIT
    url: https://github.com/eurofurence/reg-attendee-service/blob/main/LICENSE
  version: 0.1.0
servers:
  - url: /api/rest/v1
    description: localhost
  - url: /attsrv/api/rest/v1
    description: server
...
paths:
...
  /attendees/{id}/status:
    get:
      tags:
        - status
      summary: obtain the current status of an attendee
      description: Returns the current status of a single attendee. Note that this endpoint is also available to regular logged in users, this is why you only get the status field, not the full status change information. Request status history to get full information (privileged).
      operationId: getStatusById
      parameters:
        - name: id
          in: path
          description: Badge number of attendee to return
          required: true
          schema:
            type: integer
            minimum: 1
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusOnly'
        '400':
          description: Invalid ID supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Authorization required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: You do not have permission to see this attendee
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Attendee not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: An unexpected error occurred. This includes database errors. A best effort attempt is made to return details in the body.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
    post:
      tags:
        - status
      summary: request a status change
      description: |-
        Attempt a status change for a single attendee.        
        
        Depending on the transition, this may be available to a normal logged in user or admin only.
        
        - new 
          - from: approved, partially paid, paid, checked in, cancelled: admin only
        - approved
          - from: new, cancelled, partially paid, paid: admin only
        - partially paid
          - from: approved, paid, cancelled: admin only - the payment service access counts as admin
        - paid
          - from: approved, partially paid, cancelled: admin only - the payment service access counts as admin
        - checked in
          - from: paid: regdesk permission or admin
        - cancelled
          - from: new, approved: self or admin
          - from: partially paid, paid, checked in: admin
        - deleted
          - from: new, approved: admin (not possible if any payments were made for tax reasons)
        
        Note that there may also be situational limitations, such as you cannot check in an attendee unless paid in full.
        These conditions result in a 409 status to distinguish them from situations where the transition is 
        unavailable to the requesting user for permission reasons, which gives a 403.

        For detailed documentation of what the status values mean, see under Schemas/Status below.
      operationId: changeStatus
      parameters:
        - name: id
          in: path
          description: Badge number of attendee to return
          required: true
          schema:
            type: integer
            minimum: 1
            format: int64
      requestBody:
        description: The status change with an optional comment for the reason. Note, the timestamp will be ignored.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StatusChange'
        required: true
      responses:
        '204':
          description: successful operation
        '400':
          description: Invalid ID supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Authorization required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: You do not have permission to see this attendee, or you do not have permission to perform this status change at all. Note that situational limitations (e.g. cannot check in an unpaid attendee) result in 409 instead. If you get 403, you do not have permissions for this status transition, ever.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Attendee not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: The current situation does not allow this status change, even though you generally have permission to do it. Maybe you are trying to check in an attendee who hasn't fully paid, etc. See message and details fields of the error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: An unexpected error occurred. This includes database errors. A best effort attempt is made to return details in the body.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '502':
          description: The update leads to an update in the payment service which failed, or the mail service failed to send an email as part of the update.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
  /attendees/{id}/status-history:
    get:
      tags:
        - status
      summary: obtain the history of status changes of an attendee
      description: Returns the status history of a single attendee
      operationId: getStatusHistoryById
      parameters:
        - name: id
          in: path
          description: Badge number of attendee to return
          required: true
          schema:
            type: integer
            minimum: 1
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusHistory'
        '400':
          description: Invalid ID supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Authorization required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: You do not have permission to see this attendee, or you do not have privileges to see status history.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Attendee not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: An unexpected error occurred. This includes database errors. A best effort attempt is made to return details in the body.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
...
components:
  schemas:
...
    StatusOnly:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          enum:
            - new
            - approved
            - partially paid
            - paid
            - checked in
            - cancelled
            - deleted
            - waiting
          description: the status, detailed descriptions of the values see below under Schemas/Status (how do I reference this in OpenApi?)
    StatusChange:
      type: object
      required:
        - status
      properties:
        timestamp:
          type: string
          format: date-time
          description: The time at which the status change occurred. Ignored when you send us a status change and replaced with the current server time. Always present when we return a status change entry to you.
          example: 2006-01-02T15:04:05+07:00
        status:
          type: string
          enum:
            - new
            - approved
            - partially paid
            - paid
            - checked in
            - cancelled
            - deleted
            - waiting
          description: the status, detailed descriptions of the values see below under Schemas/Status (how do I reference this in OpenApi?)
        comment:
          type: string
          maxLength: 256
          description: The reason for the status change, e.g. cancel reason, or any other comment
          example: cancelled by own request
    StatusHistory:
      type: object
      required:
        - status_history
      properties:
        id:
          type: integer
          format: int64
          minimum: 1
          description: This badge number is automatically assigned during initial registration (when creating an attendee). Informational only.
          example: 10
        status_history:
          type: array
          items:
            type: object
            properties:
              schema:
                $ref: '#/components/schemas/StatusChange'
...
    Error:
      type: object
      required:
        - message
        - timestamp
        - requestid
      properties:
        timestamp:
          type: string
          format: date-time
          description: The time at which the error occurred.
          example: 2006-01-02T15:04:05+07:00
        requestid:
          type: string
          description: An internal trace id assigned to the error. Used to find logs associated with errors across our services. Display to the user as something to communicate to us with inquiries about the error.
          example: a8b7c6d5
        messsage:
          type: string
          description: |-
            A keyed description of the error. We do not write human readable text here because the user interface will be multi language.
            
            At this time, there are these values:
            - attendee.max_id.error (database error or overload)
            - attendee.parse.error (json body parse error)
            - attendee.data.invalid (field data failed to validate, see details for more information)
            - attendee.data.duplicate (duplicate registration - nickname + email + zip)
            - attendee.write.error (database error)
            - attendee.payment.error (payment service failure while updating attendee)
            - attendee.id.notfound (no such badge number in the database)
            - attendee.id.invalid (syntactically invalid badge number, must be positive integer)
            - admin.read.error (database error)
            - admin.write.error (database error)
            - admin.parse.error (json body parse error)
            - admin.data.invalid (field data failed to validate, see details for more information)
            - auth.unauthorized (token missing completely or invalid)
            - auth.forbidden (permissions missing)
            - status.read.error (database error)
            - status.write.error (database error)
            - status.mail.error (mail service failure while doing status change)
            - status.payment.error (payment service failure while doing status change)
            - status.parse.error (json body parse error)
            - status.unchanged.invalid (not actually a status change)
            - status.data.invalid (field data failed to validate, see details for more information)            
            - status.unpaid.dues (this status change is blocked because there are outstanding payments)
            - status.has.paid (this status change is impossible because there is a nonzero payment balance) 
            - status.cannot.delete (deletion is not possible, e.g. there are payments, or an invoice was issued and tax law says we have to store this data for 10 years)
            - status.use.approved (you tried to go directly to partially paid, paid, or checked in from new, cancelled, deleted - please use approved, this will automatically set (partially) paid as appropriate)
          example: attendee.data.invalid
        details:
          type: array
          items:
            type: string
          description: Optional list of additional details about the error. If available, will usually contain English language technobabble.
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: A bearer or session token obtained from the auth service (which may integrate an OIDC identity provider)
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
      description: A shared secret used for local communication (also useful for local development)